Zadanie1v2

+++ include/zad1.hpp
@@ -13,7 +13,7 @@ class Tagliatelle
   Tagliatelle(double l, double w, double r): L(l), W(w), R(r) {}
   
   //zwracającą masę mąki potrzebną do przygotowania liczby porcji równej argumentowi metody
-  double ileMaki(unsigned double P) const
+  double ileMaki(unsigned P) const
   {
     return P * L * W * (1. - R) * C;
   }
